Another large refactor:
The refactor away from Artemisd proved to be a terrible idea. Shame on me.
Refactored back into an artemisd type entity system, with a much better
design more in line with what's intended for Artemis.

Refactored animations a bit, and completed implementation. Created an
AnimationSet type, and integrated Animation and AnimationSet into
ResourceManager.